refactor: update import paths to use absolute paths feat: add authentication to protected pages fix: sign out and redirect to login page fix: change import type for UserClass in FitnessClasses and ScheduleView
feat(auth.ts): add support for NextAuth with CredentialsProvider and GoogleProvider
feat(types): add types for Location, Class, UserClass, and UserClasses
refactor(tailwind.config.js): add src/components to TailwindCSS content array
