+++ app/studio/page.tsx
@@ -1,6 +1,14 @@
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
+import { redirect } from 'next/navigation';
 import Link from 'next/link';
 
-export default function Studios() {
+export default async function Studios() {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    redirect('/login');
+  }
+
   return (
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
       Studio Page - CI/CD Test - Vercel!
