+++ src/components/Schedule/Schedule.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from 'react';
 import { supabase } from '@/lib';
-import { UserClass } from '@/types';
+import { UserClass } from '@/lib/types';
 import ScheduleView from './ScheduleView';
 
 export default function Schedule({ userId }: { userId: number}) {
