+++ src/components/Schedule/ScheduleView.tsx
@@ -11,7 +11,7 @@ import {
   format, addDays, startOfWeek, isSameDay, differenceInWeeks,
 } from 'date-fns';
 import { FitnessClasses } from '../FitnessClasses';
-import { UserClass } from '@/types';
+import { UserClass } from '@/lib/types';
 import { parseLocalDate } from '@/lib';
 
 const days = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
