+++ src/lib/index.ts
@@ -1,2 +1,3 @@
 export { default as supabase } from './supabase';
 export * from './date-fns';
+export * from './types';
